turned everything to doubles -- in order to work better with cdk representation

+++ Jmol/src/org/openscience/jmol/Animate.java
@@ -58,7 +58,7 @@ import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 import javax.swing.border.TitledBorder;
 import java.util.Vector;
-import javax.vecmath.Point3f;
+import javax.vecmath.Point3d;
 
 /**
  *  @author  Bradley A. Smith (bradley@baysmith.com)
@@ -149,8 +149,8 @@ public class Animate extends JDialog implements ActionListener,
           newCoord[1] = fromCoord[1] + (i + 1) * step[1];
           newCoord[2] = fromCoord[2] + (i + 1) * step[2];
           try {
-            extraFrames[i].addAtom(atom.getType(), (float) newCoord[0],
-                (float) newCoord[1], (float) newCoord[2]);
+            extraFrames[i].addAtom(atom.getType(), newCoord[0],
+                newCoord[1], newCoord[2]);
           } catch (Exception ex) {
             System.out.println(ex);
             ex.printStackTrace();
@@ -160,9 +160,9 @@ public class Animate extends JDialog implements ActionListener,
 
       // Linearly interpolate primitive vectors and unit cell edges
       if (fromFrame instanceof CrystalFrame) {
-	  float[][] fromRprimd = ((CrystalFrame)fromFrame).getRprimd();
-	  float[][] toRprimd = ((CrystalFrame)toFrame).getRprimd();
-	  float[][] stepRprimd = new float[3][3];
+	  double[][] fromRprimd = ((CrystalFrame)fromFrame).getRprimd();
+	  double[][] toRprimd = ((CrystalFrame)toFrame).getRprimd();
+	  double[][] stepRprimd = new double[3][3];
 	  Vector fromBoxEdges = ((CrystalFrame)fromFrame).getBoxEdges();
 	  Vector toBoxEdges = ((CrystalFrame)toFrame).getBoxEdges();
 	  Vector stepBoxEdges = new Vector(fromBoxEdges.size());
@@ -175,21 +175,21 @@ public class Animate extends JDialog implements ActionListener,
 	  }
 	  
 	  for (int i=0; i< fromBoxEdges.size(); i++) { //Box edges steps
-	    Point3f step = new Point3f();
-	    step.x = (((Point3f)toBoxEdges.elementAt(i)).x 
-		      - ((Point3f)fromBoxEdges.elementAt(i)).x) /
+	    Point3d step = new Point3d();
+	    step.x = (((Point3d)toBoxEdges.elementAt(i)).x 
+		      - ((Point3d)fromBoxEdges.elementAt(i)).x) /
 	      (numberExtraFrames + 1);
-	    step.y = (((Point3f)toBoxEdges.elementAt(i)).y 
-		      - ((Point3f)fromBoxEdges.elementAt(i)).y) /
+	    step.y = (((Point3d)toBoxEdges.elementAt(i)).y 
+		      - ((Point3d)fromBoxEdges.elementAt(i)).y) /
 	      (numberExtraFrames + 1);
-	    step.z = (((Point3f)toBoxEdges.elementAt(i)).z 
-		      - ((Point3f)fromBoxEdges.elementAt(i)).z) /
+	    step.z = (((Point3d)toBoxEdges.elementAt(i)).z 
+		      - ((Point3d)fromBoxEdges.elementAt(i)).z) /
 	      (numberExtraFrames + 1);
 	    stepBoxEdges.addElement(step);
 	  }
 
 	  for (int k = 0; k < numberExtraFrames; k++) {
-	    float[][] newRprimd = new float[3][3];
+	    double[][] newRprimd = new double[3][3];
 	    for (int i=0; i<3;i++) {
 	      for (int j=0; j<3;j++) {
 		newRprimd[i][j]=fromRprimd[i][j] + (k + 1)* stepRprimd[i][j];
@@ -197,13 +197,13 @@ public class Animate extends JDialog implements ActionListener,
 	    }
 	    Vector newBoxEdges = new Vector(stepBoxEdges.size());
 	    for (int i=0; i < stepBoxEdges.size(); i++) {
-	      Point3f newPoint = new Point3f();
-	      newPoint.x = ((Point3f)fromBoxEdges.elementAt(i)).x + (k + 1) 
-		* ((Point3f)stepBoxEdges.elementAt(i)).y;
-	      newPoint.y = ((Point3f)fromBoxEdges.elementAt(i)).y + (k + 1) 
-		* ((Point3f)stepBoxEdges.elementAt(i)).z;
-	      newPoint.z = ((Point3f)fromBoxEdges.elementAt(i)).z + (k + 1) 
-		* ((Point3f)stepBoxEdges.elementAt(i)).x;
+	      Point3d newPoint = new Point3d();
+	      newPoint.x = ((Point3d)fromBoxEdges.elementAt(i)).x + (k + 1) 
+		* ((Point3d)stepBoxEdges.elementAt(i)).y;
+	      newPoint.y = ((Point3d)fromBoxEdges.elementAt(i)).y + (k + 1) 
+		* ((Point3d)stepBoxEdges.elementAt(i)).z;
+	      newPoint.z = ((Point3d)fromBoxEdges.elementAt(i)).z + (k + 1) 
+		* ((Point3d)stepBoxEdges.elementAt(i)).x;
 	      newBoxEdges.addElement(newPoint);
 	    }
